Support filtering by execution status
I have a requirement to have different jobs depending on the status - to run
failing tests more frequently than passing tests.

This change provides checkboxes in the job configuration to select which
statuses to run. If no statuses are selected then the behavior is it is
before this change was introduced.

The reason why test cases are filtered locally is because the API seems to
only support filtering by one execution status and not multiple comma
separated statuses as the testlink-java-api documentation suggests.