Added parameters at global config:   urlResults (Url link to results in Testlink) --> used in TestLinkHelper.createReportSummary = link to the url of results in Testlink with the testplan id resolved by the job   TestLink Max. Parallel request: This defines the max number of request that can be done in parallel to TestLink. This will be used to speed up the communication with testlink. -> Parallel request=Parallel threads.
Added parameter at job config (TestLinkBuilder->config.jelly):
  keywords: Filter for use cases considering the keywords defined. If
left empty will not be used in the filter.

Added parameter at result seeker:
  keywordsExecutedFilter -> temporaly must be dissabled, it's
implemented but commented in the java code with "TODO" at the time of
the API call. 

Added parameters only in TestNGMethodNameResultSeeker:
  attachTestCasePdfReport: When enabled, the plugin will attach a "pdf"
file containing the run tests with the execution results of the use
cases in Testlink.	
  testCasesReportFolder: Folfer to find the pdf repots generated by the
tests.